healthai_property_website/backend/pyproject.toml
[project]
name = "property-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "authlib",
    "python-jose[cryptography]",
    "motor",
    "pydantic",
    "python3-saml",
    "httpx"
]
